backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
openai==1.59.7
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
flower==2.0.1
boto3==1.35.95
python-multipart==0.0.20
